App repackaging refactor
