+++ app/src/main/java/com/yusufmendes/sisterslabgraduationproject/ui/view/bag/BagFragment.kt
@@ -1,4 +1,4 @@
-package com.yusufmendes.sisterslabgraduationproject.view.bag
+package com.yusufmendes.sisterslabgraduationproject.ui.view.bag
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,7 +6,7 @@ import android.view.View
 import androidx.fragment.app.viewModels
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.yusufmendes.sisterslabgraduationproject.R
-import com.yusufmendes.sisterslabgraduationproject.adapter.BagProductAdapter
+import com.yusufmendes.sisterslabgraduationproject.ui.adapter.BagProductAdapter
 import com.yusufmendes.sisterslabgraduationproject.databinding.FragmentBagBinding
 import com.yusufmendes.sisterslabgraduationproject.util.extensions.showSnackbar
 import dagger.hilt.android.AndroidEntryPoint
