+++ app/src/main/res/layout/fragment_bag.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/screen_background"
-    tools:context=".view.bag.BagFragment">
+    tools:context=".ui.view.bag.BagFragment">
 
     <include
         android:id="@+id/bagScreenToolbar"
