+++ app/src/main/res/layout/fragment_detail.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/screen_background"
-    tools:context=".view.detail.DetailFragment">
+    tools:context=".ui.view.detail.DetailFragment">
 
     <include
         android:id="@+id/detailScreenToolbar"
